+++ flmm/PackageManager/FomodBuilder/SourceFileTree.cs
@@ -228,10 +228,13 @@ namespace Fomm.PackageManager.FomodBuilder
 			else
 			{
 				string strExtension = Path.GetExtension(p_strFile).ToLowerInvariant();
-				string strIconPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName()) + strExtension;
-				File.CreateText(strIconPath).Close();
-				imlIcons.Images.Add(strExtension, System.Drawing.Icon.ExtractAssociatedIcon(strIconPath));
-				File.Delete(strIconPath);
+				if (!imlIcons.Images.ContainsKey(strExtension))
+				{
+					string strIconPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName()) + strExtension;
+					File.CreateText(strIconPath).Close();
+					imlIcons.Images.Add(strExtension, System.Drawing.Icon.ExtractAssociatedIcon(strIconPath));
+					File.Delete(strIconPath);
+				}
 				tndFile.ImageKey = strExtension;
 				tndFile.SelectedImageKey = strExtension;
 				if (tndFile.IsArchive)
